Add configure.ac and Makefile.am to build library libMVL
Imported and adapted configure.ac and Makefile.am used with libspiro.
Currently set library C:R:A to same 1.1.0 value seen in:
https://cran.r-project.org/web/packages/RMVL/index.html
(note: linux displays C:R:A as C:A:R, other distros may show otherwise).

Set package major:minor to 0.0 since this is the first package made,
however, it may be better to set this value to start as 1.0 to sync
with RMVL value to keep numbers the same or similar.